+++ src/Support/Constants.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Convenia\Pigeon\Support;
+
+class Constants
+{
+    public const  IGNORE_PRECONDITION = 'IGNORE_PRECONDITION';
+    public const  REPLACE_ON_PRECONDITION = 'REPLACE_ON_PRECONDITION';
+    public const  THROW_PRECONDITION = 'THROW_PRECONDITION';
+}