+++ CHANGELOG.md
@@ -4,8 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Laravel 6 support
+
 ## [1.0.0]
-### Added 
+### Added
 - Added headers to publisher
 - Added `headers` to `ResolverContract`
 - Added `assertEmitted` to `Pigeon::fake()`