+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0]
 ### Added 
-- Added headers to publisher
+- Added `headers` to `Publisher`
+- Added `headers` to `ResolverContract`
 - Added `assertEmitted` to `Pigeon::fake()`
 - Added `assertConsumingEvent` to `Pigeon::fake()`
 - Added `dispatchListener` to `Pigeon::fake()`