+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Laravel 6 support
 
 ## [1.0.0]
-### Added
-- Added headers to publisher
+### Added 
+- Added `headers` to `Publisher`
+- Added `headers` to `ResolverContract`
 - Added `assertEmitted` to `Pigeon::fake()`
 - Added `assertConsumingEvent` to `Pigeon::fake()`
 - Added `dispatchListener` to `Pigeon::fake()`