+++ CHANGELOG.md
@@ -3,8 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased]
+### Fixed
+- Fixed `connection closed` sending RPC from consumer
 
-## [1.2.0]
+## [1.2.0-beta.1]
 ### Added
 - Add `application_headers` to `Pigeon::emmit` as last parameter
 - Add configurable precondition catch on queue creation